full of bugs

+++ src/main/resources/db/migration/V6__create_user_id_column_posts.sql
@@ -1,2 +1,2 @@
-ALTER TABLE posts ADD user_id int;
+ALTER TABLE posts ADD user_id bigint;
 ALTER TABLE posts ADD FOREIGN KEY (user_id) REFERENCES users(id);
